- Add AutomaticAllocationScope trait to Toy_ProgramOp (fix alloca lowering message.)
	'memref.alloca' op requires an ancestor op with AutomaticAllocationScope trait

- For foo.toy, the MLIR looked like:

I have a toy MLIR -> LLVM-IR lowering project in:https://github.com/peeterjoot/toycalculator.gitone of my test cases has the following MLIR:"builtin.module"() ({
  "toy.program"() ({
    %0 = "memref.alloca"() <{operandSegmentSizes = array<i32: 0, 0>}> : () -> memref<f64>
    "toy.declare"() <{name = "x"}> : () -> ()
    %1 = "arith.constant"() <{value = 3 : i64}> : () -> i64
    "memref.store"(%1, %0) : (i64, memref<f64>) -> ()
    "toy.assign"(%1) <{name = "x"}> : (i64) -> ()
    "toy.print"(%0) : (memref<f64>) -> ()
    "toy.return"() : () -> ()
  }) : () -> ()
}) : () -> ()

During lowering, the verify operation produced:

	'memref.store' op failed to verify that type of 'value' matches element type of 'memref'
	mlir-asm-printer: 'builtin.module' failed to verify and will be printed in generic form

This is because the alloca had f64 type, but the arith.constant did not.

There was compensation for this in NegOpLowering, and BinaryOpLowering, so the lowering worked despite the message.

To fix properly, disable that code and generate the constant as floating point instead.